feat (EBD-8):
- 1. `dockerfile` 및 `docker-compose.yaml` 파일 수정
- `dockerfile`: 최신 버전의 `spark`를 설치하도록 수정했습니다.
- `docker-compose.yaml` [트러블 슈팅 링크](https://www.notion.so/Docker-1aff0db78b5480a6af77e9c08b90590d?pvs=4#1bdf0db78b54803bb854d9ba8778d8af)
    - Airflow 관련 서비스와 Postgres 컨테이너가 동일한 Docker 네트워크에 속해 있는지 확인 후 명시적으로 네트워크 지정
    - Postgres 서비스에 네트워크 alias 추가
- 2. `insert_into_initial_processed_data.sql` 쿼리 변경
- `PRODUCT_DIM` 테이블 생성 쿼리 변경
- `brand_frequency CTE`: 각 `product_id에` 대해 브랜드(`brand`)의 발생 빈도(`brand_frequency`)를 계산하고, 해당 브랜드의 최신 `event_time을` 구한다 (빈도 중복 시 최신 데이터에 우선 순위)
- `mode_brand CTE`: 각 `product_id` 그룹 내에서 브랜드의 빈도와 최신 `event_time을` 기준으로 순위 파악
- `price_frequency CTE`: 원본 데이터(`raw_data.raw_data`)에서 각 제품(`product_id`)별, 브랜드(`brand`), 그리고 가격(`price`)의 발생 빈도와 최신 적용 시점(`event_time`)을 집계
- `mode_price CTE`: `price_frequency`에서 집계된 데이터를 기반으로, 각 제품과 브랜드 그룹별로 대표 가격(최빈값)을 결정하기 위한 순위 파악
- 최종적으로 여러개의 브랜드, 가격을 일원화 (중복 데이터 처리 및 가장 최신 값 또는 최빈값으로 대체)
- 3. `weekly_dags` 파일 생성
- 현재 진행중. 추가적으로 코드 작업 예정
- 4. 파일 이름 수정
- 파일 명의 etl → dags: ETL 작업만 수행하는 것이 아니기 때문에 dags로 일반화
- 기존에 작업했던 `initial dags` 들은 파일명 앞에 `initial`이라고 명시
