+++ Dockerfile.airflow
@@ -15,9 +15,9 @@ RUN apt-get update && apt-get install -y \
     && apt-get clean && rm -rf /var/lib/apt/lists/*
 
 # Apache Spark 설치
-RUN curl -o /tmp/spark.tgz https://dlcdn.apache.org/spark/spark-3.5.4/spark-3.5.4-bin-hadoop3.tgz \
+RUN curl -L -o /tmp/spark.tgz https://dlcdn.apache.org/spark/spark-3.5.5/spark-3.5.5-bin-hadoop3.tgz \
     && tar -xvzf /tmp/spark.tgz -C /opt/ \
-    && mv /opt/spark-3.5.4-bin-hadoop3 /opt/spark \
+    && mv /opt/spark-3.5.5-bin-hadoop3 /opt/spark \
     && rm /tmp/spark.tgz
 
 # 하둡-aws 및 aws-sdk-bundle JAR 파일 복사
