+++ dags/initial_dags/__init__.py
@@ -1,0 +1,1 @@
+from .load_to_snowflake import *
