--- dags/initial_etl/__init__.py
@@ -1,1 +1,0 @@
-from .load_to_snowflake import *
