+++ docker-compose.yaml
@@ -1,3 +1,9 @@
+# version: '3.8'
+
+networks:
+  airflow_net:
+    driver: bridge
+
 x-airflow-common: &airflow-common
   build:
     context: .
@@ -20,13 +26,14 @@ x-airflow-common: &airflow-common
     - ./data:/opt/airflow/data
     - ./spark:/opt/airflow/spark
     - ./sql:/opt/airflow/sql
-
   user: '${AIRFLOW_UID:-50000}:0'
   depends_on: &airflow-common-depends-on
     redis:
       condition: service_healthy
     postgres:
       condition: service_healthy
+  networks:
+    - airflow_net
 
 services:
   postgres:
@@ -43,6 +50,10 @@ services:
       retries: 5
       start_period: 5s
     restart: no
+    networks:
+      airflow_net:
+        aliases:
+          - postgres
 
   redis:
     image: redis:latest
@@ -55,6 +66,8 @@ services:
       retries: 50
       start_period: 30s
     restart: no
+    networks:
+      - airflow_net
 
   airflow-webserver:
     <<: *airflow-common
@@ -75,6 +88,8 @@ services:
     environment:
       <<: *airflow-common-env
       _PIP_ADDITIONAL_REQUIREMENTS: apache-airflow-providers-apache-spark boto3 pandas
+    networks:
+      - airflow_net
 
   airflow-scheduler:
     <<: *airflow-common
@@ -93,6 +108,8 @@ services:
     environment:
       <<: *airflow-common-env
       _PIP_ADDITIONAL_REQUIREMENTS: apache-airflow-providers-apache-spark boto3 pandas
+    networks:
+      - airflow_net
 
   airflow-worker:
     <<: *airflow-common
@@ -112,11 +129,15 @@ services:
       <<: *airflow-common-env
       DUMB_INIT_SETSID: '0'
       _PIP_ADDITIONAL_REQUIREMENTS: apache-airflow-providers-apache-spark boto3 pandas
+    ports:
+      - "4040:4040"   # Spark UI 포트 매핑 추가
     restart: no
     depends_on:
       <<: *airflow-common-depends-on
       airflow-init:
         condition: service_completed_successfully
+    networks:
+      - airflow_net
 
   airflow-triggerer:
     <<: *airflow-common
@@ -139,6 +160,8 @@ services:
     environment:
       <<: *airflow-common-env
       _PIP_ADDITIONAL_REQUIREMENTS: apache-airflow-providers-apache-spark boto3 pandas
+    networks:
+      - airflow_net
 
   airflow-init:
     <<: *airflow-common
@@ -210,6 +233,8 @@ services:
     user: '0:0'
     volumes:
       - ${AIRFLOW_PROJ_DIR:-.}:/sources
+    networks:
+      - airflow_net
 
   airflow-cli:
     <<: *airflow-common
@@ -226,6 +251,8 @@ services:
       <<: *airflow-common-depends-on
       airflow-init:
         condition: service_completed_successfully
+    networks:
+      - airflow_net
 
   spark-master:
     build:
@@ -247,7 +274,7 @@ services:
       - ./spark:/opt/spark
       - ./spark/conf:/opt/bitnami/spark/conf
       - ./data:/opt/airflow/data
-    user: "1001:1001"  # 👈 spark 사용자가 실행하도록 설정
+    user: "1001:1001"
     depends_on:
       - redis
       - postgres
@@ -258,6 +285,8 @@ services:
       retries: 5
       start_period: 30s
     restart: no
+    networks:
+      - airflow_net
 
   spark-worker:
     build:
@@ -267,7 +296,7 @@ services:
       - SPARK_MODE=worker
       - SPARK_MASTER_URL=spark://spark-master:7077
       - SPARK_WORKER_CORES=2
-      - SPARK_WORKER_MEMORY=2G
+      - SPARK_WORKER_MEMORY=4G
       - SPARK_RPC_AUTHENTICATION_ENABLED=no
       - SPARK_RPC_ENCRYPTION_ENABLED=no
       - SPARK_LOCAL_STORAGE_ENCRYPTION_ENABLED=no
@@ -287,6 +316,8 @@ services:
       retries: 5
       start_period: 30s
     restart: no
+    networks:
+      - airflow_net
 
 volumes:
   postgres-db-volume:
