+++ spark/conf/spark-env.sh
@@ -48,3 +48,50 @@ LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
 LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
 LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
 LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
+LD_PRELOAD=/opt/bitnami/common/lib/libnss_wrapper.so
